Gateway/AI/Qwen TTS/requirements.txt
# Generated from pyproject.toml dependencies
transformers==4.57.3
accelerate==1.12.0
gradio
librosa
torchaudio
soundfile
sox
onnxruntime
einops
